pyproject.toml
[project]
name = "quiz-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn>=0.35.0",
]
